Added partial impl of jpa repository

+++ rest/src/main/java/gxc/rcalc/rest/repository/RiskIndexRepository.java
@@ -1,12 +1,14 @@
 package gxc.rcalc.rest.repository;
 
+import gxc.rcalc.rest.entity.RiskIndex;
+
 import java.util.Collection;
 
 public interface RiskIndexRepository {
 	
 	RiskIndex createOrUpdate(RiskIndex index);
 	
-	RiskIndex findById(Integer id);
+	RiskIndex findById(Long id);
 	
 	Collection<RiskIndex> findByCompanyName(String companyName);
